- Moved exception class to correct package  - Added preliminary checking for coursestudent existence in coursestudent creation through rest service
